finish 01 and 02, 02 with some hiccups
I individually set some elements to display: inline-block so that they
could create some space between the edges of the card. This solution was
very manual: some of the elements were funky and the width of some
elements had to be eyeballed. The better solution (which was in the
solution file) was setting the padding of the parent of ALL these
elements (the <div> with .card class) to 8px. The children of the .card
class are the box contents OF the div, so they could easily have space
between them just by adding the 8px padding assigned. I also forgot
to add the properties of the button, but it was easily fixed.