Finish 05
Spent a great amount of time figuring out how to assemble the text
on the right of the '!' mark. Figured out that the warning header and
the 'X' button were on the same line, and the rest of the boxes were
below it. My solution is slightly off from the assigned solution but I
don't think it's a big idea. The '!' Icon was shrinked when setting the
display of .modal to flex, so to counteract that it was simply putting
.icon {flex-shrink: 0;}.